change urls in docs to use example.net/.com
Prior to this commit, the docs and vmfloaty.yml.example file used
a variety of live domains (mycompany.net, alternate.net). This commit
changes those to example.net and example.com which are IANA-managed
Reserved Domains. This will reduce the chance that any credentials
could be harvested if a user executes the example commands without
changing the domains.